rename exchanges dir to providers
