+++ contracts/dex/src/providers/exchange_resolver.rs
@@ -4,19 +4,19 @@ use crate::DexCommand;
 
 // Supported exchanges on Juno
 #[cfg(feature = "juno")]
-pub use crate::exchanges::{
+pub use crate::providers::{
     junoswap::{JunoSwap, JUNOSWAP},
     wyndex::{WynDex, WYNDEX},
 };
 
 #[cfg(feature = "terra")]
-pub use crate::exchanges::terraswap::{Terraswap, TERRASWAP};
+pub use crate::providers::terraswap::{Terraswap, TERRASWAP};
 
 #[cfg(feature = "terra")]
-pub use crate::exchanges::astroport::{Astroport, ASTROPORT};
+pub use crate::providers::astroport::{Astroport, ASTROPORT};
 
 #[cfg(any(feature = "juno", feature = "osmosis"))]
-pub use crate::exchanges::osmosis::{Osmosis, OSMOSIS};
+pub use crate::providers::osmosis::{Osmosis, OSMOSIS};
 
 pub(crate) fn identify_exchange(value: &str) -> Result<&'static dyn Identify, DexError> {
     match value {
